Igniter.Code.Function (igniter v0.3.29)

Utilities for working with functions.

Functions

append_argument(zipper, value)

@spec append_argument(Sourceror.Zipper.t(), any()) ::
  {:ok, Sourceror.Zipper.t()} | :error

Appends an argument to a function call, leaving the zipper at the function call's node.

argument_equals?(zipper, index, term)

@spec argument_equals?(Sourceror.Zipper.t(), integer(), any()) :: boolean()

Checks if the provided function call (in a Zipper) has an argument that equals term at index.

argument_matches_pattern?(zipper, index, pattern)
(macro)

Returns true if the argument at the provided index exists and matches the provided pattern

Note: to check for argument equality, use argument_equals?/3 instead.

argument_matches_predicate?(zipper, index, func)

@spec argument_matches_predicate?(
  Sourceror.Zipper.t(),
  non_neg_integer(),
  (Sourceror.Zipper.t() ->
     boolean())
) :: boolean()

Returns true if the argument at the given index matches the provided predicate

function_call?(zipper)

@spec function_call?(Sourceror.Zipper.t()) :: boolean()

Returns true if the node is a function call

function_call?(zipper, name, arity \\ :any)

@spec function_call?(
  Sourceror.Zipper.t(),
  atom() | {module(), atom()},
  arity :: integer() | :any
) ::
  boolean()

Returns true if the node is a function call of the given name

If an atom is provided, it only matches functions in the form of function(name).

If an {module, atom} is provided, it matches functions called on the given module, taking into account any imports or aliases.

move_to_def(zipper, fun, arity)

@spec move_to_def(
  Sourceror.Zipper.t(),
  fun :: atom(),
  arity :: integer() | [integer()]
) ::
  {:ok, Sourceror.Zipper.t()} | :error

move_to_defp(zipper, fun, arity)

@spec move_to_defp(
  Sourceror.Zipper.t(),
  fun :: atom(),
  arity :: integer() | [integer()]
) ::
  {:ok, Sourceror.Zipper.t()} | :error

move_to_function_call(zipper, name, arity, predicate \\ fn _ -> true end)

Moves to a function call by the given name and arity, matching the given predicate, in the current or lower scope

move_to_function_call_in_current_scope(zipper, name, arity, predicate \\ fn _ -> true end)

Moves to a function call by the given name and arity, matching the given predicate, in the current scope

move_to_nth_argument(zipper, index)

@spec move_to_nth_argument(
  Sourceror.Zipper.t(),
  non_neg_integer()
) :: {:ok, Sourceror.Zipper.t()} | :error

Moves to the nth argument of a function call.

update_nth_argument(zipper, index, func)

@spec update_nth_argument(
  Sourceror.Zipper.t(),
  non_neg_integer(),
  (Sourceror.Zipper.t() -> {:ok, Sourceror.Zipper.t()} | :error)
) :: {:ok, Sourceror.Zipper.t()} | :error

Updates the nth argument of a function call, leaving the zipper at the function call's node.