Igniter.Code.Module (igniter v0.3.29)

Utilities for working with Elixir modules

Functions

create_module(igniter, module_name, contents, opts \\ [])

Creates a new file & module in its appropriate location.

find_all_matching_modules(igniter, predicate)

@spec find_all_matching_modules(igniter :: Igniter.t(), (module(),
                                                   Sourceror.Zipper.t() ->
                                                     boolean())) ::
  {Igniter.t(), [module()]}

find_and_update_module(igniter, module_name, updater)

@spec find_and_update_module(Igniter.t(), module(), (Sourceror.Zipper.t() ->
                                                 {:ok, Sourceror.Zipper.t()}
                                                 | :error)) ::
  {:ok, Igniter.t()} | {:error, Igniter.t()}

Finds a module and updates its contents. Returns {:error, igniter} if the module could not be found. Do not discard this igniter.

find_and_update_module!(igniter, module_name, updater)

find_and_update_or_create_module(igniter, module_name, contents, updater, opts \\ [])

Finds a module and updates its contents wherever it is.

If the module does not yet exist, it is created with the provided contents. In that case, the path is determined with Igniter.Code.Module.proper_location/2, but may optionally be overwritten with options below.

Options

:path - Path where to create the module, relative to the project root. Default: nil (uses :kind to determine the path).

find_module(igniter, module_name)

@spec find_module(Igniter.t(), module()) ::
  {:ok, {Igniter.t(), Rewrite.Source.t(), Sourceror.Zipper.t()}}
  | {:error, Igniter.t()}

Finds a module, returning a new igniter, and the source and zipper location. This new igniter should not be discarded.

In general, you should not use the returned source and zipper to update the module, instead, use this to interrogate the contents or source in some way, and then call find_and_update_module/3 with a function to perform an update.

module?(zipper)

module_exists?(igniter, module_name)

Checks if a module is defined somewhere in the project. The returned igniter should not be discarded.

module_name(suffix)

This function is deprecated. Use `module_name/2` instead..

@spec module_name(String.t()) :: module()

Given a suffix, returns a module name with the prefix of the current project.

module_name(igniter, suffix)

@spec module_name(Igniter.t(), String.t()) :: module()

Given a suffix, returns a module name with the prefix of the current project.

module_name_prefix()

This function is deprecated. Use `module_name_prefix/1` instead.

@spec module_name_prefix() :: module()

The module name prefix based on the mix project's module name

module_name_prefix(igniter)

@spec module_name_prefix(Igniter.t()) :: module()

The module name prefix based on the mix project's module name

move_to_def(zipper, fun, arity)

This function is deprecated. Use `Igniter.Code.Function.move_to_def/3` instead.

move_to_defmodule(zipper)

@spec move_to_defmodule(Sourceror.Zipper.t()) :: {:ok, Sourceror.Zipper.t()} | :error

Moves the zipper to a defmodule call

move_to_defmodule(zipper, module)

@spec move_to_defmodule(Sourceror.Zipper.t(), module()) ::
  {:ok, Sourceror.Zipper.t()} | :error

Moves the zipper to a specific defmodule call

move_to_defp(zipper, fun, arity)

This function is deprecated. Use `Igniter.Code.Function.move_to_defp/3` instead.

move_to_module_using(zipper, one_of_modules)

@spec move_to_module_using(Sourceror.Zipper.t(), module() | [module()]) ::
  {:ok, Sourceror.Zipper.t()} | :error

Moves the zipper to the body of a module that uses the provided module (or one of the provided modules).

move_to_use(zipper, module)

Moves the zipper to the use statement for a provided module.

move_to_using(zipper, module)

This function is deprecated. Use `move_to_use/2` instead..

parse(module_name)

@spec parse(String.t()) :: module()

Parses a string into a module name

proper_location(module_name, source_folder \\ "lib")

@spec proper_location(module(), source_folder :: String.t()) :: Path.t()

Returns the idiomatic file location for a given module, starting with "lib/".

Examples:

iex> Igniter.Code.Module.proper_location(MyApp.Hello) "lib/my_app/hello.ex"

proper_test_location(module_name)

@spec proper_test_location(module()) :: Path.t()

Returns the test file location for a given module, according to mix test expectations, starting with "test/" and ending with "_test.exs".

Examples:

iex> Igniter.Code.Module.proper_test_location(MyApp.Hello) "test/my_app/hello_test.exs"

iex> Igniter.Code.Module.proper_test_location(MyApp.HelloTest) "test/my_app/hello_test.exs"

proper_test_support_location(module_name)

@spec proper_test_support_location(module()) :: Path.t()

Returns the test support location for a given module, starting with "test/support/" and dropping the module name prefix in the path.

Examples:

iex> Igniter.Code.Module.proper_test_support_location(MyApp.DataCase) "test/support/data_case.ex"