Igniter.Libs.Phoenix (igniter v0.3.29)

Codemods & utilities for working with Phoenix

Functions

add_pipeline(igniter, name, contents, opts \\ [])

add_scope(igniter, route, contents, opts \\ [])

endpoints_for_router(igniter, router)

@spec endpoints_for_router(igniter :: Igniter.t(), router :: module()) ::
  {Igniter.t(), [module()]}

list_routers(igniter)

select_router(igniter, label \\ "Which router should be modified?")

web_module(igniter)

@spec web_module(Igniter.t()) :: module()

Returns the web module name for the current app

web_module_name()

This function is deprecated. Use `web_module/0` instead..

@spec web_module_name() :: module()

Returns the web module name for the current app

web_module_name(suffix)

This function is deprecated. Use `web_module_name/1` instead..

@spec web_module_name(String.t()) :: module()

Generates a module name that lives in the Web directory of the current app.

web_module_name(igniter, suffix)

@spec web_module_name(Igniter.t(), String.t()) :: module()